Backport of [JDK-8302191](https://bugs.openjdk.java.net/browse/JDK-8302191). Applies cleanly. It is already committed in trunk.

```
build/linux-x86_64-server-fastdebug/jdk/bin/java -jar ~/azul/jmh-jdk-microbenchmarks-git/micros-jdk8/target/micros-jdk8-1.0-SNAPSHOT.jar org.openjdk.bench.java.lang.FloatDoubleRem -i 2 -r 2 -wi 2 -f 2
unpatched:
Benchmark                            Mode  Cnt    Score    Error  Units
FloatDoubleRem.calcDoubleJava        avgt    4  100.167 ±  9.248  ns/op
FloatDoubleRem.calcFloatJava         avgt    4   97.162 ± 55.984  ns/op
FloatDoubleRem.cornercaseDoubleJava  avgt    4    7.325 ±  2.565  ns/op
FloatDoubleRem.cornercaseFloatJava   avgt    4    7.363 ±  0.981  ns/op
patched:
Benchmark                            Mode  Cnt   Score   Error  Units
FloatDoubleRem.calcDoubleJava        avgt    4  31.459 ± 2.992  ns/op
FloatDoubleRem.calcFloatJava         avgt    4  31.526 ± 5.581  ns/op
FloatDoubleRem.cornercaseDoubleJava  avgt    4  10.144 ± 1.094  ns/op
FloatDoubleRem.cornercaseFloatJava   avgt    4  10.214 ± 0.932  ns/op
```